backend/requirements.txt
asgiref==3.8.1
<<<<<<< HEAD
Django==4.2.18
django-cors-headers==4.6.0
djangorestframework==3.15.2
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.12.2
=======
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==4.2.18
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
jwt==1.3.1
mysqlclient==2.2.7
pycparser==2.22
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
>>>>>>> f28d7c64b9f09a5318a3482598d0f4d5ce9e3a7f
